+++ scripts/count-strict-null-check-errors.cjs
@@ -1,5 +1,3 @@
-/* eslint-disable @typescript-eslint/no-var-requires */
-/* eslint-disable no-console */
 const fs = require("fs");
 const readline = require("readline");