+++ src/extensions/views/InstallCustomExtension/components/InstallCustomExtensionFromForm/InstallCustomExtensionFromForm.tsx
@@ -1,7 +1,9 @@
 import { Savebar } from "@dashboard/components/Savebar";
+import { ExtensionManifestValidator } from "@dashboard/extensions/domain/extension-manifest-validator";
 import { headerTitles, messages } from "@dashboard/extensions/messages";
 import { useAutoSubmit } from "@dashboard/utils/hook-form/auto-submit";
 import { Box } from "@saleor/macaw-ui-next";
+import { useMemo } from "react";
 import {
   Control,
   UseFormGetValues,
@@ -18,6 +20,8 @@ import { ExtensionInstallFormData } from "../../types";
 import { InstallSectionData } from "../InstallSectionData/InstallSectionData";
 import { InstallTopNav } from "../InstallTopNav";
 import { ManifestUrlForm } from "../ManifestUrlForm/ManifestUrlForm";
+
+const manifestValidator = new ExtensionManifestValidator();
 
 export const InstallCustomExtensionFromForm = ({
   control,
@@ -51,6 +55,18 @@ export const InstallCustomExtensionFromForm = ({
     manifest,
   });
 
+  const issues = useMemo(() => {
+    const manifestValidation = manifestValidator.validateAppManifest(manifest);
+
+    return "issues" in manifestValidation ? manifestValidation.issues : undefined;
+  }, [manifest]);
+
+  /**
+   * Prevent installation if validation fails. In the future we can change errors to warnings, but first we need to add special handling
+   * to Dashboard, e.g. render "warning" near widget that widget is broken.
+   */
+  const canInstall = !issues && manifest;
+
   return (
     <>
       <InstallTopNav title={intl.formatMessage(headerTitles.addCustomExtensionManifest)} />
@@ -71,13 +87,14 @@ export const InstallCustomExtensionFromForm = ({
           manifest={manifest}
           lastFetchedManifestUrl={lastFetchedManifestUrl}
           control={control}
+          issues={issues}
         />
       </Box>
       <Savebar>
         <Savebar.Spacer />
         <Savebar.CancelButton href={previousPagePath} />
         <Savebar.ConfirmButton
-          disabled={!manifest}
+          disabled={!canInstall}
           transitionState={isSubmittingInstallation ? "loading" : "default"}
           onClick={() => submitInstallApp()}
         >