+++ src/products/components/ProductUpdatePage/types.ts
@@ -92,7 +92,7 @@ export interface UseProductUpdateFormOutput
   formErrors: FormErrors<ProductUpdateSubmitData>;
 }
 
-type UseProductUpdateFormRenderProps = Omit<UseProductUpdateFormOutput, "datagrid" | "richText">;
+type UseProductUpdateFormRenderProps = Omit<UseProductUpdateFormOutput, "datagrid">;
 
 export interface UseProductUpdateFormOpts
   extends Record<"categories" | "collections" | "taxClasses", Option[]> {